cc430 code ready for github - readme file still needs to be updated

+++ core/net/glossy.h
@@ -139,7 +139,7 @@ uint8_t glossy_get_relay_cnt_first_rx(void);
 /**
  * @brief get the packet error rate/ratio in percentages
  */
-uint8_t glossy_get_per(void);
+uint16_t glossy_get_per(void);
